Message Boxes & Splash
Made the message boxes and splash screens a bit more elegent

+++ WHA Revit Launcher.ahk
@@ -54,8 +54,7 @@ IfNotExist %iniPathLocal%
 	FileCopy, %A_ScriptDir%\%programName%.ini, %iniPathLocal%
 	If ErrorLevel ; If copying the file goes wrong, notify and exit
 	{
-		MsgBox, 16, %programName%, We were unable to properly initialize %programName%.  Please see %bimGuy% for additional information.
-		ExitApp
+		PrettyMsg("We were unable to properly initialize " . programName . ".  Please see " . bimGuy . " for additional information.", "exit")
 	}
 	; As long as we are setting things up, make sure the program starts when the user logs including
 	; This can be turned off in settings
@@ -68,8 +67,7 @@ IfNotExist %iniPathLocal%
 iniLocal := class_EasyIni(iniPathLocal) 
 If (!iniLocal.Settings.iniPathCentral) ;Check if settings file is good
 {
-	MsgBox, 16, %programName%, A file named "%iniPathLocal%" either could not be found in the same directory as this executable or it was manually altered.  This program will now exit.`n`nPlease contact your friendly BIM Coordinator.
-	ExitApp
+	PrettyMsg("A file named """ . iniPathLocal . """ either could not be found in the same directory as this executable or it was manually altered.  This program will now exit.`n`nPlease contact your friendly BIM Coordinator.", "exit")
 }
 
 ; Load up the global project list
@@ -260,7 +258,7 @@ Else
 		; Check if project still exists
 		if !(favVersion%A_Index%)
 		{
-			MsgBox, % "Uh oh, The project:`n`n" . fav%A_Index%Name . "`n`ncould not be found in the Global Project List. It will be removed from your quick launch menu.`n`nYou may want to try adding it again. Please see " . bimGuy . " should this problem persist."
+			PrettyMsg("Uh oh, The project:`n`n" . fav%A_Index%Name . "`n`ncould not be found in the Global Project List. It will be removed from your quick launch menu.`n`nYou may want to try adding it again. Please see " . bimGuy . " should this problem persist.")
 			iniLocal.DeleteSection(favList%A_Index%)
 			iniLocal.Save()
 			LogMe("Favorite", "Remove", favList%A_Index%, "No longer in global list")
@@ -313,7 +311,7 @@ return
 
 ; If the server is not available, check again when the user requests
 TrayReset:
-MsgBox, 53, %programName%, The centralized list of projects could not be found.`nThis may be due to the network not being available.  Please check the network and try again.  Consult %bimGuy% should this problem persist.
+PrettyMsg("The centralized list of projects could not be found.`nThis may be due to the network not being available.  Please check the network and try again.  Consult " . bimGuy . " should this problem persist.")
 IfMsgBox Retry
 	ReloadMe()
 LogMe("MenuTray", "No Network")
@@ -606,8 +604,8 @@ RemoveProjectRemove:
 RowNumber := 0
 RowNumber := LV_GetNext(RowNumber)
 LV_GetText(rowText, rowNumber, 2)
-MsgBox, 33, Remove Project?, % "Are you sure you want to remove the following project from your list?`n`n" . iniLocal [favList%RowNumber%].Name
-IfMsgBox OK
+pMessage := PrettyMsg("Are you sure you want to remove the following project from your list?`n`n" . iniLocal [favList%RowNumber%].Name, "question")
+if (pMessage == "Yes")
 {
 	iniLocal.DeleteSection(favList%RowNumber%)
 	iniLocal.Save()
@@ -615,6 +613,8 @@ IfMsgBox OK
 	LogMe("Favorite", "Remove", favList%RowNumber%)
 	ReloadMe()
 }
+else if (pMessage == "No")
+	Gui, RemoveProject:Destroy
 return
 
 
@@ -641,12 +641,12 @@ newFavName := iniCentral [rowText].NameShort
 
 if instr(favList, newFavSection)
 {
-	MsgBox, 64, %programName%, %newFavNumber% %newFavName% is already in your list of projects
+	PrettyMsg(newFavNumber . " " . newFavName . " is already in your list of projects")
 }
 else
 {
-	MsgBox, 33, %programName%, You are sure you want to add the following project to your list?`n`n%newFavNumber% - %newFavName%
-	IfMsgBox OK
+	pMessage := PrettyMsg("Are you sure you want to add the following project to your list?`n`n" . newFavNumber . " - " . newFavName, "question")
+	if (pMessage == "Yes")
 	{
 		iniLocal.AddSection(newFavSection, "ProjectID", rowText)
 		iniLocal.AddKey(newFavSection, "Name", newFavNumber . " " . newFavName)
@@ -657,6 +657,8 @@ else
 		LogMe("Favorite", "Add", rowText, newFavNumber, newFavName, "Detach 0", "Workset 0")
 		ReloadMe()
 	}
+	else if (pMessage == "No")
+		Gui, AddProject:Destroy
 }
 return
 ; ### End of Add/Remove Quick Launch Menu ###
@@ -747,7 +749,7 @@ Gui, Settings:Add, Button, w75 xs+185 ys+45 gSettingsSubLocalDefault, Default
 Gui, Settings:Font, s12 cBlack, Arial
 Gui, Settings:Add, Text, xm yp+50 section,
 Gui, Settings:Font, s18, Arial
-Gui, Settings:Add, Text, xs+100 ys-5, Log File
+Gui, Settings:Add, Text, xs+100 ys-5, Log Files
 Gui, Settings:Font, s9 c%guiColor1%, Arial
 Gui, Settings:Add, Text, xs+100 ys+25 w400 vlogLocation, %localLog%
 Gui, Settings:Add, Button, w75 xs+100 ys+45 gSettingsSubLog, View
@@ -774,7 +776,6 @@ return
 
 
 SettingsSubDetach:
-; msgBox, % !detachDefault
 iniLocal.Settings.Detach := !detachDefault
 iniLocal.Save()
 detachDefault := iniLocal.Settings.Detach
@@ -786,7 +787,6 @@ return
 
 
 SettingsSubWorkset:
-; msgBox, % !detachDefault
 iniLocal.Settings.Workset := !worksetDefault
 iniLocal.Save()
 worksetDefault := iniLocal.Settings.Workset
@@ -817,7 +817,7 @@ IfNotExist, %LinkFile%
 	FileCreateShortcut, %A_ScriptFullPath%, %LinkFile% 
 	If ErrorLevel
 	{
-		MsgBox, 16, %programName%, There was a problem creating a shortcut in your startup folder, %A_Startup%.  Please contact %bimGuy% for additional information.
+		PrettyMsg("There was a problem creating a shortcut in your startup folder, " . A_Startup . ".  Please contact %bimGuy% for additional information.")
 		LogMe("Error", "Shortcut", A_ScriptFullPath, LinkFile)
 	}
 }
@@ -826,7 +826,7 @@ Else
 	FileDelete, %LinkFile%
 	If ErrorLevel
 	{
-		MsgBox, 1, %programName%, There was a problem removing the startup shortcut from your startup folder, %A_Startup%. Please contact %bimGuy% for additional information.
+		PrettyMsg("There was a problem removing the startup shortcut from your startup folder, " . A_Startup . ". Please contact %bimGuy% for additional information.")
 	}
 }
 IfExist %linkFile%
@@ -841,6 +841,7 @@ Return
 
 SettingsSubLocal:
 ;Allows user to set location where local files will be saved. It also give the option to set it back to the default location.
+PrettyMsg("Please make sure to set a folder on your local computer. Revit doesn't handle local files located on network drives well.")
 FileSelectFolder, selectedFolder, , 3, Choose a location for your local files to reside:
 If ErrorLevel
 	Return
@@ -855,17 +856,17 @@ Return
 
 SettingsSubLocalDefault:
 IfEqual, localFolder, %A_MyDocuments%\Revit
-	MsgBox, 16, %programName%, The local file save location is currently set to the default location.
+	PrettyMsg("The local file save location is currently set to the default location.")
 Else
 {
-	MsgBox, 17, %programName%, This will return the local file save location to the default setting.  Are you sure you would like to proceed?
-	IfMsgBox, Cancel
-		Return
-	localFolder = %A_MyDocuments%\Revit
-	iniLocal.Settings.LocalFolder := localFolder
-	iniLocal.save()
-	GuiControl, , localLocation, %localFolder%
-	LogMe("Settings", "localLocation", localFolder, "Default")
+	if PrettyMsg("This will return the local file save location to the default setting.  Are you sure you would like to proceed?", "question", 2) == "Yes"
+	{
+		localFolder = %A_MyDocuments%\Revit
+		iniLocal.Settings.LocalFolder := localFolder
+		iniLocal.save()
+		GuiControl, Settings:, localLocation, %localFolder%
+		LogMe("Settings", "localLocation", localFolder, "Default")
+	}
 }
 Return
 
@@ -905,13 +906,16 @@ iniCentral := class_EasyIni(iniPathCentral)
 ; Set a variable to use the Manage Add menu for both adding and editing
 maEdit := 0
 ; Create the add project menu
+manageWidth := 680
+manageButtonWidth := (manageWidth - (15*2)) / 3
+manageButtonLoc := manageButtonWidth + 15
 Gui, Manage:New,, %programName%
 Gui, Manage:Font, s24 , Arial
-Gui, Manage:Add, Text, center w680, Select a project to manage:
+Gui, Manage:Add, Text, center w%manageWidth%, Select a project to manage:
 Gui, Manage:Font, s18 , Arial
-Gui, Manage:Add, Button, w680 gManageProjectAdd, &Add a New Project
+Gui, Manage:Add, Button, w%manageWidth% gManageProjectAdd, &Add a New Project
 Gui, Manage:Font, s10 , Arial
-Gui, Manage:Add, ListView, AltSubmit r15 w680 gManageProjectList -Multi, Launcher ID|Project Number|Name|Version
+Gui, Manage:Add, ListView, AltSubmit r15 w%manageWidth% gManageProjectList -Multi, Launcher ID|Project Number|Name|Version
 GoSub, ManageListUpdate
 LV_ModifyCol(1, 100)
 LV_ModifyCol(2, 100)
@@ -919,9 +923,9 @@ LV_ModifyCol(3, 400)
 LV_ModifyCol(4, 59)
 Gui, Manage:Font, s10 c%guiColor1%, Arial
 Gui, Manage:Font, s18 cBlack, Arial
-Gui, Manage:Add, Button, w200 yp+350 xm gManageProjectEdit, &Edit Project
-Gui, Manage:Add, Button, wp xp+240 gManageProjectRemove, &Remove Project
-Gui, Manage:Add, Button, wp xp+240 Default gManageGuiClose, &Exit
+Gui, Manage:Add, Button, w%manageButtonWidth% yp+350 xm gManageProjectEdit, &Edit Project
+Gui, Manage:Add, Button, wp xp+%manageButtonLoc% gManageProjectRemove, &Remove Project
+Gui, Manage:Add, Button, wp xp+%manageButtonLoc% Default gManageGuiClose, &Exit
 GuiControl, Manage:Disable, Button2
 GuiControl, Manage:Disable, Button3
 Gui, Manage:Show
@@ -1034,7 +1038,7 @@ Gui, ManageAdd:Submit, NoHide ;Don't hide the window until the data has been che
 ;Check that version is in right format
 If !(RegExMatch(maVersion, "P)^20\d\d$"))
 {
-	MsgBox, 64, %programName%, You have not entered a valid Revit Version.  We are looking for the year of the release only.  For example, if the project was last saved in Revit 2014 or Revit Architecture 2014, enter "2014" for the version.
+	PrettyMsg("You have not entered a valid Revit Version.  We are looking for the year of the release only.  For example, if the project was last saved in Revit 2014 or Revit Architecture 2014, enter ""2014"" for the version.")
 	Gui, Font, s10 cRed, Arial
 	GuiControl, Font, maVersionText
 	Return
@@ -1052,7 +1056,8 @@ Gui, ManageAdd:Destroy
 ;Create new Section in the global ini file
 If maEdit
 {
-	MsgBox, 49, %programName%, Are you sure you want to update this project?`n`nNumber: %maProject%`nName: %maName%`nShort Name: %maShort%`nVersion: %maVersion%`nCentral: %maCentral%`nProject Folder: %maWorkingFolder%
+	if !PrettyMsg("Are you sure you want to update this project?`n`nNumber: " . maProject . "`nName: " . maName . "`nShort Name: " . maShort . "`nVersion: " . maVersion . "`nCentral: " . maCentral . "`nProject Folder: " . maWorkingFolder, "question", 2)
+		return
 	iniCentral.DeleteSection(rowText)
 	LogMe("ManageList", "Pre-edit", rowText, oldProject, oldName, oldShort, oldVersion, oldCentral, oldWorkingFolder)
 	LogMe("ManageList", "Edit", maLauncherID, maProject, maName, maShort, maVersion, maCentral, maWorkingFolder)
@@ -1096,14 +1101,14 @@ rowNumber := LV_GetNext(RowNumber)
 LV_GetText(rowText, rowNumber)
 delPNumber := iniCentral [rowText].Number
 delPName := iniCentral [rowText].Name
-MsgBox, 49, %programName%, Are you sure you want to delete project:`n%delPNumber% %delPName%`n`nThis will effect all users in the office.
-IfMsgBox, Cancel
-	Return
-iniCentral.DeleteSection(rowText)
-LogMe("ManageList", "Remove", rowText, delPNumber, delPName)
-GoSub, ManageListUpdate
-iniCentral.Save()
-; newFavID := centralSections%RowNumber%
+if PrettyMsg("Are you sure you want to delete project:`n" . delPNumber . " " . delPName . "`n`nThis will effect all users in the office.", "question", 2)
+{
+	iniCentral.DeleteSection(rowText)
+	LogMe("ManageList", "Remove", rowText, delPNumber, delPName)
+	GoSub, ManageListUpdate
+	iniCentral.Save()
+	; newFavID := centralSections%RowNumber%
+}
 Return
 
 
@@ -1300,7 +1305,7 @@ If exploreLaunch
 If !detach
 {
 	launchStatus = Copying local...
-	GuiControl,, LaunchText, %launchStatus%
+	GuiControl, Launch:, LaunchText, %launchStatus%
 	DebugMe("CreateDir")
 	;Check if Directory Exists
 	IfNotExist, %projectFolder%
@@ -1309,8 +1314,7 @@ If !detach
 		FileCreateDir, %projectFolder% ;Create a directory if it doesn't
 		If Errorlevel
 		{
-			MsgBox, 16, Could not create directory, A directory structure could not be created in %localFolder%.
-			ExitApp
+			PrettyMsg("A directory structure could not be created in " . localFolder ".`nPlease contact " . bimGuy, "exit")
 		}
 	}
 
@@ -1360,15 +1364,19 @@ If !detach
 	FileCopy, %pCentral%, %localPath%, 1
 	if ErrorLevel ;Check to make sure everything copied correctly
 	{
-		MsgBox, 16, Local Creation Error, The local file could not be copied to your computer.  Please see your BIM manager for additional information.
-		ExitApp
 		LogMe("Launcher", "Error", "LocalCreate", projectID, localPath, pCentral)
+		PrettyMsg("The local file could not be copied to your computer.  Please see your BIM manager for additional information.", "exit")
 	}
 }
 
 DebugMe("Launch")
-launchStatus = Launching Revit %pVersion%
-GuiControl,, LaunchText, %launchStatus%
+if Detach
+	launchStatus = Detaching a Revit %pVersion% Model
+else if Specify
+	launchStatus = Launching Revit %pVersion% with Specify
+else
+	launchStatus = Launching a Revit %pVersion% Model
+GuiControl, Launch:, LaunchText, %launchStatus%
 ;Open Revit with the correct version
 If monitorPath ;Open the worksharing monitor if it exists
 	Run, %monitorPath%
@@ -1380,7 +1388,7 @@ IfWinExist, ^%revitTitle% ;Check to see if Revit is already running
 Else
 {
 	Run, %revitPath%, Max, %localPath%
-	WinWait, ^%revitTitle%
+	WinWait, ^%revitTitle% - \[Recent Files\]
 	WinMaximize
 }
 WinActivate
@@ -1389,15 +1397,15 @@ WinWait, Project Not Saved Recently,, 1
 if !Errorlevel
 {
 	Gui, Launch:Destroy
-	MsgBox, 48, %programName%, Please save your current project/family before launching a new one.
+	PrettyMsg("Please save your current project/family before launching a new one.")
 	LogMe("Launcher", "Error", "Project Not Saved Dialog", projectID, revitPath, revitTitle, localPath)
 	ReloadMe("noshow")
 }
-WinWait, Open,, 15
+WinWait, Open,, 30
 If ErrorLevel
 {
 	Gui, Launch:Destroy
-	MsgBox, 48, %programName%, There seems to be an issue launching your project. Check Revit for any opened dailog boxes and try launching again.`n`nThanks.
+	PrettyMsg("There seems to be an issue launching your project. Check Revit for any opened dailog boxes and try launching again.`n`nThanks.")
 	LogMe("Launcher", "Error", "OpenWait", projectID, revitPath, revitTitle, localPath)
 	ReloadMe("noshow")
 }
@@ -1406,15 +1414,13 @@ WinGet, openID, ID, Open
 ControlGet, fileHwnd, Hwnd,, Edit1, ahk_id %openID%
 ControlGet, folderHwnd, Hwnd,, SysListView321, ahk_id %openID%
 ControlGet, openHwnd, Hwnd,, Button1, ahk_id %openID%
-; MsgBox Open Window: %openID%`nFile name: %fileHwnd%`nFolderView:%folderHwnd%`nOpen Button: %openHwnd%
-If detach
+if detach
 	fName := pCentral
 Else
 	fName := localPath
 If workset
 {
 	SplitPath, fName, fName, fPath
-	; MsgBox, 1, Workset Flow, Workset: %workset% File: %fPath%
 	Sleep 300
 	Control, EditPaste, %fPath%, , ahk_id %fileHwnd%
 	ControlClick, Button1, Open,, L, 2, NA
@@ -1424,7 +1430,6 @@ If workset
 }
 Else
 {
-	; msgbox, 1, Regular Flow, Workset: %workset% File: %fName%
 	Control, EditPaste, %fName%, Edit1, Open
 }
 If detach
@@ -1469,31 +1474,32 @@ ReloadMe("noshow")
 Return
 
 LaunchSplash:
+launchWidth := 500
+SysGet, pMonitor, MonitorPrimary
+SysGet, pMonitor, MonitorWorkArea, %pMonitor%
 Gui, Launch:New
-If debug
-	Gui, Launch:-Caption ; AlwaysOnTop 
-Else
-	Gui, Launch:-Caption
-Gui, Launch:Margin, 50, 50
+Gui, Launch:-Caption AlwaysOnTop
+Gui, Launch:Margin, 50, 25
 Gui, Launch:Color, DCDCDC
-whaLogo(700, "Launch")
-Gui, Launch:Font, cBlack s36, Arial
-Gui, Launch:Add, Text,xm yp+150 center w700 vLaunchText, Launching...
-Gui, Launch:Font, c%guiColor1% s18, Arial
-Gui, Launch:Add, Text,yp+70 center w700, %pNumber% %pName%
-Gui, Launch:Font, c%guiColor2% s24, Arial
-if detach or workset
-{
-	Gui, Launch:Add, Text, yp+70 w700 vlaunchSub center,
-	if detach & workset
-		GuiControl, Launch:Text, launchSub, Detach and Specify
-	else if workset
-		GuiControl, Launch:Text, launchSub, Specify Worksets
-	else
-		GuiControl, Launch:Text, launchSub, Detach
-}
+whaLogo(launchWidth, "Launch")
+Gui, Launch:Font, cBlack s24, Arial
+Gui, Launch:Add, Text, xm yp+45 center w%launchWidth% vLaunchText, Launching...
+Gui, Launch:Font, c%guiColor1% s12, Arial
+Gui, Launch:Add, Text, yp+40 center w%launchWidth% vLaunchSub, %pNumber% %pName%
+Gui, Launch:Font, c%guiColor2% s18, Arial
+launchHeight := 150
+launchY := pMonitorBottom - launchHeight
+;~ if detach or workset
+;~ {
+	;~ if detach & workset
+		;~ GuiControl, Launch:Text, launchSub, Detach and Specify %pNumber% %pName%
+	;~ else if workset
+		;~ GuiControl, Launch:Text, launchSub, Specify Worksets %pNumber% %pName%
+	;~ else
+		;~ GuiControl, Launch:Text, launchSub, Detach %pNumber% %pName%
+;~ }
 
-Gui, Launch:Show
+Gui, Launch:Show, y%launchY% h%launchHeight%
 return
 ; ### End of Main Routine ###
 
@@ -1570,7 +1576,6 @@ ReloadMe(sx = "show")
 MouseCleanClick(x, y)
 {
 	MouseGetPos, px, py
-	; MsgBox, %x%,%y%`n%px%,%py%
 	Click %x%, %y%
 	Click %px%, %py%, 0
 }
@@ -1586,12 +1591,6 @@ DebugMe(debugText)
 		ReloadMe()
 	}
 }
-
-BetterMsgBox(message)
-{
-
-}
-Return
 
 LogMe(category, params*)
 {
